// car-gal/interfaces.ts
export interface dynCssVariables {
    id: string;
    value: string;
    childValues?: dynCssVariables[]
}

export interface IcGElementStyleObject {
    id?: string | string[];
    values?: string[][];
    childValues?: IcGElementStyleObject[];
}